Подключение работает
Изменил структуру таблиц: тип id и Like
Связал с помощью библиотеки dio с Flutter, пока что без nginx

+++ Application/Entities/Step.cs
@@ -2,9 +2,9 @@
 {
     public class Step
     {
-        public long StepId { get; set; }
+        public int StepId { get; set; }
         public string Description { get; set; }
 
-        public long RecipeId { get; set; }
+        public int RecipeId { get; set; }
     }
 }
